new routes working: establishment, product, address, redemptionCode

+++ src/controllers/addressController.js
@@ -1,14 +1,5 @@
 const addressService = require("../services/addressService");
 
-const getAddressController = async (req, res) => {
-  try {
-    const result = await addressService.getAddress();
-    res.status(200).json({ message: "Adressess fetched successfully", address: result  });
-  } catch (error) {
-    res.status(500).json({ error: "Internal server error" });
-  }
-};
-
 const createAddressController = async (req, res) => {
   try {
     const result = await addressService.createAddress(req.body);
@@ -19,8 +10,12 @@ const createAddressController = async (req, res) => {
 };
 
 const updateAddressController = async (req, res) => {
+  const { addressId } = req.params;
+  const userType = req.user.idTipoUtilizador;
   try {
-    const { addressId } = req.params;
+    if (userType !== 1) {
+      return res.status(403).json({ error: "Access denied. Admins only." });
+    }
     await addressService.updateAddress(addressId, req.body);
     res.status(200).json({ message: "Address updated successfully"});
   } catch (error) {
@@ -30,7 +25,11 @@ const updateAddressController = async (req, res) => {
 
 const deleteAddressController = async (req, res) => {
   const { addressId } = req.params;
+  const userType = req.user.idTipoUtilizador;
   try {
+    if (userType !== 1) {
+      return res.status(403).json({ error: "Access denied. Admins only." });
+    }
     await addressService.deleteAddress(addressId);
     res.status(200).json({message: "Address deleted successfully"});
   } catch (error) {
@@ -38,4 +37,4 @@ const deleteAddressController = async (req, res) => {
   }
 };
 
-module.exports = { getAddressController, createAddressController, updateAddressController, deleteAddressController };
+module.exports = {createAddressController, updateAddressController, deleteAddressController };
